api/requirements.txt
fastapi
uvicorn
websockets
pydantic
python-multipart
docker
openai
motor
python-dotenv
